requirements.txt
mitreattack-python
networkx
pyvis
textwrap